Resolves divergence: local (1 commit + uncommitted work, committed as wip
26ae65d) vs origin/master (10 commits). Merged along two independent axes:

- Deps/imports (origin canonical): monstertruck 0.3 meta-crate migration
  (monstertruck::geometry/modeling/step), bevy 0.18, wgpu 29,
  bevy_panorbit_camera, and origin's monstertruck.rs coarse-merged shell
  export rewrite.
- far public API (local kept, more idiomatic): refine_adaptive takes
  Option<&[Index]> (None == origin's &[] behaviorally), TopologyDescriptor
  ::creases -> Result<Self>. Aligned origin's wholesale-taken call sites
  (monstertruck.rs test, bevy.rs, topology_descriptor doctest) to these.
- bfr::SurfaceFactory::new approx params -> i32 to match origin's
  monstertruck.rs callers (body casts via `as _`, no behavior change).

Dropped: local planar-surface-drop filter in monstertruck.rs (recoverable
from 26ae65d); origin's STEP-gen rewrite supersedes it.

Verified: cargo check --features monstertruck --all-targets passes; clippy
clean (default features); all non-STEP tests pass. 7 STEP-export golden
tests fail PRE-EXISTING on origin (goldens at 2eb8883 predate origin's
monstertruck.rs rewrite) -- regenerate with
RUST_UPDATE_EXPECTED_TEST_RESULTS=1 if origin's current output is authoritative.

+++ opensubdiv-petite/src/bfr.rs
@@ -29,8 +29,8 @@ impl SurfaceFactory {
     /// smooth and sharp features.
     pub fn new(
         refiner: &crate::far::TopologyRefiner,
-        approx_smooth: usize,
-        approx_sharp: usize,
+        approx_smooth: i32,
+        approx_sharp: i32,
     ) -> Result<Self, Error> {
         unsafe {
             let ptr = sys::bfr::surface_factory::Bfr_SurfaceFactory_Create(
@@ -185,7 +185,7 @@ impl Surface {
 }
 
 #[cfg(feature = "monstertruck")]
-use monstertruck_geometry::prelude::{BsplineSurface, KnotVector, Point3};
+use monstertruck::geometry::prelude::{BsplineSurface, KnotVector, Point3};
 
 #[cfg(feature = "monstertruck")]
 impl SurfaceFactory {
